requirements.txt
anyio==3.6.2
certifi==2022.12.7
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
python-multipart==0.0.6
PyYAML==6.0
sniffio==1.3.0
sqladmin==0.10.3
SQLAlchemy==2.0.11
starlette==0.26.1
typing_extensions==4.5.0
WTForms==3.0.1
fastapi~=0.95.1
pydantic~=1.10.7
uvicorn==0.22.0
aiomonitor~=0.4.5